refactor: core/material を整理しパッケージ化、後方互換ガードを削除
- __init_subclass__ の旧 API 検出ガード 11 種を削除 (hardening_type / implicit_stress / 旧シグネチャ等)
- 基底 MaterialModel.isotropic_C と _default_stress_update を dead code として削除
- 各サブクラスで重複していた elastic_stiffness を基底に lift
- _default_stress_residual を公開 default_stress_residual に inline
- material.py を base.py / stress_states.py へ分割しパッケージ化
- 既存 import パスは __init__.py の re-export で維持
