+++ tests/unit/test_material_bases.py
@@ -541,77 +541,6 @@ def test_implicit_state_without_state_residual_raises():
                 pass
 
 
-def test_hardening_type_attribute_raises():
-    """Declaring hardening_type must raise TypeError with migration hint."""
-    with pytest.raises(TypeError, match="hardening_type has been removed"):
-        class Bad(MaterialModel3D):
-            hardening_type = "augmented"
-            param_names = []
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def state_residual(self, state_new, dlambda, state_n, state_trial):
-                return {}
-
-
-def test_hardening_type_reduced_hint():
-    """hardening_type='reduced' gives hint to remove the attribute."""
-    from manforge.core.state import Explicit
-    with pytest.raises(TypeError, match="implicit_state_names=\\[\\] is the default"):
-        class Bad(MaterialModel3D):
-            hardening_type = "reduced"
-            param_names = []
-            ep = Explicit(shape=())
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def update_state(self, dlambda, state_n, state_trial):
-                return [self.ep(state_n["ep"] + dlambda)]
-
-
-def test_list_state_names_raises():
-    """Declaring non-empty list state_names raises TypeError with migration hint."""
-    with pytest.raises(TypeError, match="list-based state_names has been removed"):
-        class Bad(MaterialModel3D):
-            param_names = []
-            state_names = ["ep"]
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def update_state(self, dlambda, state_n, state_trial):
-                return [state_n["ep"] + dlambda]
-
-
-def test_list_implicit_state_names_raises():
-    """Declaring non-empty list implicit_state_names raises TypeError with migration hint."""
-    from manforge.core.state import Implicit
-    with pytest.raises(TypeError, match="list-based implicit_state_names has been removed"):
-        class Bad(MaterialModel3D):
-            param_names = []
-            implicit_state_names = ["alpha"]
-            alpha = Implicit(shape=NTENS)
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def state_residual(self, state_new, dlambda, state_n, state_trial):
-                return {"alpha": anp.array(0.0)}
-
 
 def test_mixed_implicit_explicit_requires_both_methods():
     """Partial implicit (some explicit, some implicit) requires both methods."""
@@ -671,110 +600,6 @@ def test_no_state_fields_no_methods_needed():
     assert OKStateless.state_names == ["stress"]
     assert OKStateless.implicit_state_names == []
 
-
-def test_implicit_stress_true_raises():
-    """Declaring implicit_stress=True must raise TypeError with migration hint."""
-    from manforge.core.state import Implicit
-    with pytest.raises(TypeError, match="implicit_stress has been removed"):
-        class Bad(MaterialModel3D):
-            implicit_stress = True
-            param_names = []
-            alpha = Implicit(shape=NTENS)
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def state_residual(self, state_new, dlambda, state_n, state_trial):
-                return [self.alpha(anp.array(0.0))]
-
-
-def test_implicit_stress_false_raises():
-    """Declaring implicit_stress=False must raise TypeError with migration hint."""
-    with pytest.raises(TypeError, match="implicit_stress has been removed"):
-        class Bad(MaterialModel3D):
-            implicit_stress = False
-            param_names = []
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def update_state(self, dlambda, state_n, state_trial):
-                return []
-
-
-def test_old_yield_function_signature_raises():
-    """yield_function(self, stress, state) must raise TypeError."""
-    with pytest.raises(TypeError, match="yield_function.*is no longer accepted"):
-        class Bad(MaterialModel3D):
-            param_names = []
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, stress, state):
-                pass
-
-            def update_state(self, dlambda, state_n, state_trial):
-                return []
-
-
-def test_old_update_state_signature_raises():
-    """update_state(self, dlambda, stress, state) must raise TypeError."""
-    with pytest.raises(TypeError, match="update_state.*is no longer accepted"):
-        class Bad(MaterialModel3D):
-            param_names = []
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def update_state(self, dlambda, stress, state):
-                return []
-
-
-def test_old_state_residual_signature_raises():
-    """state_residual(self, state_new, dlambda, stress, state_n) must raise TypeError."""
-    from manforge.core.state import Implicit
-    with pytest.raises(TypeError, match="state_residual.*is no longer accepted"):
-        class Bad(MaterialModel3D):
-            param_names = []
-            alpha = Implicit(shape=NTENS)
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def state_residual(self, state_new, dlambda, stress, state_n):
-                return []
-
-
-def test_stress_residual_override_raises():
-    """Overriding stress_residual() must raise TypeError with migration hint."""
-    with pytest.raises(TypeError, match="stress_residual.*override has been removed"):
-        class Bad(MaterialModel3D):
-            param_names = []
-
-            def elastic_stiffness(self, state=None):
-                pass
-
-            def yield_function(self, state):
-                pass
-
-            def update_state(self, dlambda, state_n, state_trial):
-                return []
-
-            def stress_residual(self, stress, dlambda, state, stress_trial, state_n):
-                pass
 
 
 # ---------------------------------------------------------------------------
