+++ tests/unit/test_state_fields.py
@@ -382,45 +382,3 @@ def test_make_state_missing_raises():
         model.make_state(alpha=np.zeros(6))
 
 
-# ---------------------------------------------------------------------------
-# Migration errors
-# ---------------------------------------------------------------------------
-
-def test_list_state_names_non_empty_raises():
-    with pytest.raises(TypeError, match="list-based state_names has been removed"):
-        class Bad(MaterialModel3D):
-            param_names = []
-            state_names = ["ep"]
-
-            def yield_function(self, state):
-                return anp.array(0.0)
-
-            def update_state(self, dlambda, state_n, state_trial):
-                return []
-
-
-def test_list_implicit_state_names_non_empty_raises():
-    with pytest.raises(TypeError, match="list-based implicit_state_names has been removed"):
-        class Bad(MaterialModel3D):
-            param_names = []
-            implicit_state_names = ["alpha"]
-            alpha = Implicit(shape=NTENS)
-
-            def yield_function(self, state):
-                return anp.array(0.0)
-
-            def state_residual(self, state_new, dlambda, state_n, state_trial):
-                return [self.alpha(state_new["alpha"] - state_n["alpha"])]
-
-
-def test_empty_list_state_names_allowed():
-    """state_names=[] is a no-op (backwards compat); stress is still auto-attached."""
-    class OK(MaterialModel3D):
-        param_names = []
-        state_names = []
-
-        def yield_function(self, state):
-            return anp.array(0.0)
-
-    # stress is always auto-attached even when no user fields are declared
-    assert OK.state_names == ["stress"]
